The crash bug happens when Blinker.begin then BTmenu.begin are both run, in that order.
If the order of those two is reversed, all setup completes
without crash.

If the BTmenu.loop is activated, crash will happen
regardless of the setup load order, and regardless
of whether Blinker.loop is running or not.

The bottom line is that the Blinker and the BTmenu
loop are not yet able to run at the same time
without crashing.