Reverted switch statements to use raw 'byt', comparing it with single-quoted characters (which are interpreted as ascii code). This is how it was before impementing selected_menu_item. Note that we're keeping the selecte_menu_item, as its useful for other functionality as well.
This is only for SerialMenu. Storage uses literal integers
in switch statement.

Comparing raw byte with single-quoted ascii chr is good,
because it gives us more range of what we can compare.

Disabled end-of-admin reboot, now just goes to run mode,
as it did awhile ago. This needs to be tested for potential
security issues when transitioning from admin to run
under different global var conditions.