Fixed bug in RFID timing by always using uint32_t for time-related data.
Now able to handle large tag-timeout and reader-high-cycle values,
which means you can set the tag-timeout to 8+hours and the 'aging'
stage beggining just shortly before the end of that. This would
give you the practical equivalent to a "tag-in-once" system.
Under this scenario, the 'aging' stage should have an audible
beeper or at least a bright flashing light to warn the user
that a re-tag-in is required. This scenario is useful for setups
where the tag is not always available to the reader, like with extremely
short-range readers.