Debuged SerialMenu::prompt().
Implemented Settings::displaySetting(int) to return
a formatted string continaing a single setting's name-value pair.

Settings::displaySetting(int) returns a poiter to a char[]
and is intended for UI display, to be used in a for-loop,
iterating over the length of the Settings array.

Call this from anywhere. Use SETTINGS_SIZE and other globals
to wrap calls to Settings::displaySetting(int).