Refactored Storage and its subclassing in Settings to use simple function overrloads when subclassed from Storage.
This mostly eliminates the need for class (static) variables in
Storage and its derived classes and does not require templates.

The variables that were static (and whos base required function templates)
are now passed to the base functions as arguments. The subclassed
functions wrap the base functions and pass in the class-specific data.

Testing first compiled build that satisfies design requirements.