Added mediumBeep() to Led, which takes number of cycles. This is called whenever admin mode is activated.
Added code to Menu to listen to reader periodically and force
exit of admin mode as soon as any tag is read. Currently doesn't work.

Changed Uptime() and PreLog() to use static-const-char, instead
of heap variable (new char[22]), for internal buffer. This is
intended to reduce the liability of calling functions to free()
the returned heap memory. It is not working correctly...
It results in old data in buffer being written to log.

This build crashes the ARD after awhile being up.