Converting serial ports to use SerialPort class for all but the 'begin()' command.
This commit works with the Pro Mini containing the Uno bootloader.
BT serial is on the main HW serial pins, but it still doesn't work
for uploads. Using hard-wired serial port works well for uploads.

Added 555-based schmidt trigger to power. It drops ARD reset pin
to LOW when main power is disconnected briefly. A ~5000mf cap downstream
of a diode from main power keeps the ARD and BT modules running,
while the 555 is powered from raw 5v, before the diode/cap. This allows
the 555 and ARD reset pin to go LOW, while the ARD itself and the BT module
continue to run. This will trigger a reset-with-bootloader on the
ARD Uno bootloader (Optiboot).