Refactored use of HW and SW serial ports to use custom SerialPort class.
Debugged changes.

Disabled Menu::M2 and SW serial port, just to demonstrate
that it can be done. We should re-enable them if possible,
in case we/users want to attach a 2nd serial monitor.
This could solve the issue of how to manage BT AT commands,
while BT is on the HW serial port.